Add round robin check on interapp communication.
