+++ middle-app/Dockerfile
@@ -1,0 +1,15 @@
+FROM node:argon
+
+RUN mkdir -p /usr/src/app
+WORKDIR /usr/src/app
+
+COPY package.json /usr/src/app
+RUN npm install
+
+COPY . /usr/src/app
+
+EXPOSE 3333
+
+CMD ["npm", "start"]
+
+
