+++ nginx/nginx.conf
@@ -10,7 +10,7 @@ server {
     # (workaround for a limitation of nginx)
     # see https://www.ruby-forum.com/topic/4407628
 
-    set $alias "mynodeapp:3333";
+    set $alias "mymiddleapp:3333";
     proxy_pass http://$alias/;
 
     proxy_redirect off;
